remove dependency to eclipse
packaging a 10MiB library into the resulting hpi just to make use of
immutable lists and sets seems a bit overkill.
Such things are available from plain Java or guava (comes with core).

Also remove dependency to servlet-api which is provided by core

Still open is to replace the dependency to jackson with the api plugin
as it currently doesn't include the dataformat-csv package (PR opened)